Replace the SonarQube server with the analyser in the build
The server could never reach the workflow that needed it. SonarQube Community
Build has no branch analysis: every run landed on the project's single `main`
branch whatever the working tree was on, which made a per-branch analysis
meaningless and two concurrent ones destructive. So `feature-oneshot` was told
not to run it, `milestone-review` ran it once a milestone, and CI never ran it
at all — an analysis needs `secrets.SONAR_TOKEN` and Test-NoCloudCredentials.ps1
fails the build on any `secrets.` expression, so the one place every change
passes through was the one place the analysis was absent.

SonarAnalyzer.CSharp is the same rule engine, shipped as a Roslyn analyser. As a
GlobalPackageReference it runs inside every `dotnet build`: every branch, every
agent's checkout, every CI job, no server, no token, no scan step. Warnings are
errors repo-wide and the S-rules are not exempted, so a finding fails the build
in the phase that introduced it rather than landing in a report somebody has to
go and read.

Deleted, and recoverable from here: docker-compose.sonarqube.yml,
infra/sonarqube/, build/ci/Invoke-SonarAnalysis.ps1, SonarQube.Analysis.xml.
Invoke-Verification.ps1 loses its Sonar stage and -ReuseCoverageForCrap, which
existed only to reuse that stage's coverage; it is two stages and 300 lines
lighter. What went with the server: the dashboard, the quality gate, issue
history, duplication density and SonarQube's own coverage view. Coverage and
complexity are the CRAP stage's job.

THE FIRST RUN: 492 findings. 188 fixed, 15 carry a one-line #pragma with its
reason, 289 are covered by 18 .editorconfig entries — each one verified against
the code before the rule was switched off, because the default profile collides
with conventions this repository already settled. The documentation-constant
convention (a named constant instead of a design-doc citation in a comment)
reads as 37 dead fields. Prose comments read as commented-out code. Exact float
comparison IS the determinism contract. `res://` paths read as URIs.
`ThrowIfNull(item, nameof(equipped))` — blaming the argument the caller actually
passed — reads as a mistake.

Three findings would have made the code WORSE if obeyed, which is the argument
for reading a rule before following it:

  - Dropping the `_ =` discard on a fire-and-forget reintroduced CS4014, an
    error here.
  - Removing a "dead" store to `reached` would have made a failed profile-open
    report itself as a failed session-open: the catch handlers read it, and the
    await between the two stores can throw.
  - Collapsing `Hash64.Of(new Hash64Argument[]{…})` to its elements would have
    bound both sides of a test to the same overload, turning a real assertion
    into x == x.

Worth fixing, from the 188: two iterator methods whose ThrowIfNull was deferred
to the first MoveNext instead of firing on the call, a duplicated test whose
name promised an assertion its body never made, an undisposed
CancellationTokenSource on the client's root node, and an integer division
compared against a double.

Verified on a clean --no-incremental build: 0 warnings, 0 errors. 9,242 unit
tests, 6 architecture tests, and the vendor-package gate all pass.

⚠️ An incremental build emits no diagnostics for an up-to-date project, so a
stale `dotnet build` can look clean. Changed projects always re-analyse;
--no-incremental forces the full sweep.

+++ tests/SlayIdleRepeat.Core.Tests/Rules/Combat/BattleClockTests.cs
@@ -47,7 +47,7 @@ public sealed class BattleClockTests
 
             if (tick % CombatLog.TicksPerSecond == 0)
             {
-                seconds.ShouldBe(tick / CombatLog.TicksPerSecond);
+                seconds.ShouldBe(tick / (double)CombatLog.TicksPerSecond);
             }
         }
     }
